Update backfill required_rubygems_version task to required_ruby_version
We started storing required_ruby_version since 20 March, 2014. At
the time it was only used in UI and considered not as important to
backfill.
rubygems pre installation checks uses gemspec to verify ruby requirements,
and bundle install fails when it tries to install a version with
incompatible requirements.
Now that bundler uses required_ruby_version in dependency resolution
it would be nicer if older versions reported their required ruby
version on the compact index endpoint.

440314 versions match the `where` criteria.